+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This release contains security and bug fixes.
 - libheif: wrong memcpy size: [commit](https://github.com/strukturag/libheif/commit/11ffeffadd980f9f96019fe180fc1e81827e3790)
 - libheif: rare situation with alpha channel: [libheif issue](https://github.com/strukturag/libheif/issues/495)
 - libheif: do not pad to 16x16 pixels for AVIF: [commit](https://github.com/strukturag/libheif/commit/ec1dc464dc08517ecef7b675043886ec727eadb2)
+- libheif: scaling of images in some cases:  [commit](https://github.com/strukturag/libheif/commit/0cd461e18b99d018f9adef731eec928781078afb)
 
 ## [0.6.0 - 2022-08-06]